+++ examples/Ex201_PSC_tensorGrid.jl
@@ -41,12 +41,21 @@ function main(;
 
     ## primary data for I-V scan protocol
     scanrate = 0.4 * V / s
-    number_tsteps = 31
     endVoltage = voltageAcceptor # bias goes until the given voltage at acceptor boundary
-
     ## with fixed timestep sizes we can calculate the times a priori
     tend = endVoltage / scanrate
-    tvalues = range(0, stop = tend, length = number_tsteps)
+
+    ## Define scan protocol function
+    function linearScanProtocol(t)
+        return if t == Inf
+            0.0
+        else
+            scanrate * t
+        end
+    end
+
+    ## Apply zero voltage on left boundary and a linear scan protocol on right boundary
+    contactVoltageFunction = [zeroVoltage, linearScanProtocol]
 
     if test == false
         println("*** done\n")
@@ -127,7 +136,7 @@ function main(;
     ################################################################################
 
     ## Initialize Data instance and fill in data
-    data = Data(grid, p.numberOfCarriers)
+    data = Data(grid, p.numberOfCarriers, contactVoltageFunction = contactVoltageFunction)
 
     ## Possible choices: Stationary, Transient
     data.modelType = Transient
@@ -231,9 +240,17 @@ function main(;
 
     ################################################################################
     if test == false
-        println("I-V Measurement Loop")
+        println("I-V Measurement")
     end
     ################################################################################
+
+    control.Δt = 5.0e-2
+    control.Δt_grow = 1.03
+    ## calculation of solution
+    sol = ChargeTransport.solve(ctsys, inival = inival, times = (0.0, tend), control = control)
+
+    tvalues = sol.t
+    number_tsteps = length(tvalues)
 
     ## for saving I-V data
     IV = zeros(0) # for IV values
@@ -241,27 +258,19 @@ function main(;
 
     for istep in 2:number_tsteps
 
-        t = tvalues[istep]       # Actual time
-        Δu = t * scanrate         # Applied voltage
-        Δt = t - tvalues[istep - 1] # Time step size
-
-        ## Apply new voltage; set non equilibrium boundary conditions
-        set_contact!(ctsys, p.bregionAcceptor, Δu = Δu)
-
-        if test == false
-            println("time value: t = $(t) s")
-        end
-
-        solution = solve(ctsys, inival = inival, control = control, tstep = Δt)
+        Δt = t - tvalues[istep - 1]  # Time step size
+        inival = sol.u[istep - 1]
+        solution = sol.u[istep]
 
         ## get I-V data
         current = get_current_val(ctsys, solution, inival, Δt)
-
         push!(IV, current)
-        push!(biasValues, Δu)
 
         inival = solution
+
     end # time loop
+
+    biasValues = contactVoltageFunction[p.bregionAcceptor].(tvalues)
 
     if plotting
         Plotter.figure()
@@ -279,7 +288,7 @@ function main(;
         Plotter.zlabel("potential [V]")
         ################
         Plotter.figure()
-        Plotter.plot(biasValues, IV .* (cm)^2 / height, label = "", linewidth = 3, marker = "o")
+        Plotter.plot(biasValues[2:end], IV .* (cm)^2 / height, label = "", linewidth = 3, marker = "o")
         Plotter.grid()
         Plotter.ylabel("total current [A]") #
         Plotter.xlabel("Applied Voltage [V]")
@@ -309,7 +318,7 @@ function main(;
 end #  main
 
 function test()
-    testval = -0.5815947701306688; testvalvacancyEnergyCalculation = -0.5823076281125769
+    testval = -0.5816008029666527; testvalvacancyEnergyCalculation = -0.582313421829256
     return main(test = true) ≈ testval && main(test = true, vacancyEnergyCalculation = false) ≈ testvalvacancyEnergyCalculation
 end